@@ -1,9 +1,9 @@
 [project]
 name = "pygame-ce"
-version = "2.5.6.dev1"
+version = "2.5.7.dev1"
 description = "Python Game Development"
 readme = "README.rst" # for long description
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = {text = "LGPL v2.1"}
 authors = [{name = "A community project"}]
 classifiers = [
@@ -14,7 +14,6 @@ classifiers = [
     "Programming Language :: Cython",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -55,9 +54,9 @@ pygame_ce = 'pygame.__briefcase.pygame_ce:PygameCEGuiBootstrap'
 [build-system]
 requires = [
   "meson-python<=0.18.0",
-  "meson<=1.8.2",
-  "ninja<=1.12.1",
-  "cython<=3.1.2",
+  "meson<=1.10.0",
+  "ninja<=1.13.0",
+  "cython<=3.2.4",
   "sphinx<=8.2.3",
   "astroid<4.0.0",
   "sphinx-autoapi<=3.6.0",
@@ -77,7 +76,7 @@ install = ['--tags=runtime,python-runtime,pg-tag']
 # dependencies. Here is where uv comes into the picture. It is an "installer" like pip,
 # but faster. It has been observed to save a couple of minutes of CI time.
 build-frontend = "build[uv]"
-build = "cp3{9,10,11,12,13,14}-* pp3{10,11}-*"
+build = "cp3{10,11,12,13,14}-* pp311-*"
 skip = ["*-musllinux_*", "cp31?t-*"]
 # build[uv] is verbose by default, so below flag is not needed here
 # build-verbosity = 3
@@ -97,6 +96,21 @@ setup-args = [
     "-Dvsenv=true",  # same effect as passing --vsenv
     "-Derror_docs_missing=true",
 ]
+
+[tool.cibuildwheel.pyodide]
+build = "cp313-*"  # build only for the latest python version.
+
+# EMSDK path is hardcoded here, and has to be manually kept updated with updates
+# to cibuildwheel and/or pyodide.
+before-build = """
+sed -i 's/var SUPPORT_LONGJMP *= *[^;]*;/var SUPPORT_LONGJMP = "wasm";/' \
+  /home/runner/.cache/cibuildwheel/emsdk-4.0.9/emsdk-4.0.9/upstream/emscripten/src/settings.js &&
+embuilder --pic --force build \
+  sdl2 libhtml5 sdl2_ttf 'sdl2_mixer:formats=ogg,mp3,mod,mid' \
+  'sdl2_image:formats=bmp,gif,jpg,lbm,pcx,png,pnm,qoi,svg,tga,xcf,xpm,xv'
+"""
+test-command = ""  # TODO: figure out how to test
+
 
 [tool.ruff]
 exclude = [
