@@ -1,13 +1,13 @@
 from abc import ABC, abstractmethod
-from typing import Literal, Optional, Union
+from typing import Literal
 
 from pygame.surface import Surface
 from pygame.typing import IntPoint
 
 def get_backends() -> list[str]: ...
-def init(backend: Optional[str] = None) -> None: ...
+def init(backend: str | None = None) -> None: ...
 def quit() -> None: ...
-def list_cameras() -> list[str]: ...
+def list_cameras() -> list[str] | list[int]: ...
 def colorspace(
     surface: Surface, color: Literal["YUV", "HSV"], dest_surface: Surface = ..., /
 ) -> Surface: ...
@@ -24,7 +24,7 @@ class AbstractCamera(ABC):
     @abstractmethod
     def query_image(self) -> bool: ...
     @abstractmethod
-    def get_image(self, dest_surf: Optional[Surface] = None) -> Surface: ...
+    def get_image(self, dest_surf: Surface | None = None) -> Surface: ...
     @abstractmethod
     def get_raw(self) -> bytes: ...
     # set_controls and get_controls are not a part of the AbstractCamera ABC,
@@ -34,7 +34,7 @@ class AbstractCamera(ABC):
 class Camera(AbstractCamera):
     def __init__(
         self,
-        device: Union[str, int] = 0,
+        device: str | int,
         size: IntPoint = (640, 480),
         format: str = "RGB",
     ) -> None: ...
@@ -49,5 +49,5 @@ class Camera(AbstractCamera):
     ) -> tuple[bool, bool, int]: ...
     def get_size(self) -> tuple[int, int]: ...
     def query_image(self) -> bool: ...
-    def get_image(self, surface: Optional[Surface] = None) -> Surface: ...
+    def get_image(self, surface: Surface | None = None) -> Surface: ...
     def get_raw(self) -> bytes: ...
