@@ -50,6 +50,13 @@
 #include "stdbool.h"
 
 #if SDL_VERSION_ATLEAST(3, 0, 0)
+
+#include "include/SDL_gesture.h"
+
+#define SDL_DOLLARGESTURE GESTURE_DOLLARGESTURE
+#define SDL_DOLLARRECORD GESTURE_DOLLARRECORD
+#define SDL_MULTIGESTURE GESTURE_MULTIGESTURE
+
 #define PG_ShowCursor SDL_ShowCursor
 #define PG_HideCursor SDL_HideCursor
 #define PG_CursorVisible SDL_CursorVisible
@@ -67,12 +74,6 @@
 #define PG_AUDIO_ALLOW_FORMAT_CHANGE 0
 #define PG_AUDIO_ALLOW_CHANNELS_CHANGE 0
 #define PG_AUDIO_ALLOW_ANY_CHANGE 0
-
-// Todo: deal with multigesture.. See
-// https://github.com/pygame-community/pygame-ce/issues/2420
-#define PG_MULTIGESTURE 0
-
-#define PG_JOYBALLMOTION 0
 
 #define PG_CreateSurface SDL_CreateSurface
 #define PG_CreateSurfaceFrom SDL_CreateSurfaceFrom
@@ -191,6 +192,8 @@ PG_GetSurfaceFormat(SDL_Surface *surf)
 
 #define PG_GetSurfaceClipRect SDL_GetSurfaceClipRect
 
+#define PG_GL_SetSwapInterval SDL_GL_SetSwapInterval
+
 #else /* ~SDL_VERSION_ATLEAST(3, 0, 0)*/
 #define PG_ShowCursor() SDL_ShowCursor(SDL_ENABLE)
 #define PG_HideCursor() SDL_ShowCursor(SDL_DISABLE)
@@ -207,10 +210,6 @@ PG_GetSurfaceFormat(SDL_Surface *surf)
 #define PG_AUDIO_ALLOW_FORMAT_CHANGE SDL_AUDIO_ALLOW_FORMAT_CHANGE
 #define PG_AUDIO_ALLOW_CHANNELS_CHANGE SDL_AUDIO_ALLOW_CHANNELS_CHANGE
 #define PG_AUDIO_ALLOW_ANY_CHANGE SDL_AUDIO_ALLOW_ANY_CHANGE
-
-#define PG_MULTIGESTURE SDL_MULTIGESTURE
-
-#define PG_JOYBALLMOTION SDL_JOYBALLMOTION
 
 #define PG_CreateSurface(width, height, format) \
     SDL_CreateRGBSurfaceWithFormat(0, width, height, 0, format)
@@ -403,6 +402,12 @@ PG_GetSurfaceClipRect(SDL_Surface *surface, SDL_Rect *rect)
     *rect = surface->clip_rect;
     return true;
 }
+
+static inline bool
+PG_GL_SetSwapInterval(int interval)
+{
+    return SDL_GL_SetSwapInterval(interval) == 0;
+}
 #endif
 
 /* DictProxy is useful for event posting with an arbitrary dict. Maintains
@@ -450,6 +455,40 @@ typedef enum {
     PGM_BUTTON_KEEP = 0x80
 } PygameMouseFlags;
 
+#if SDL_VERSION_ATLEAST(3, 0, 0)
+typedef enum {
+    PGE_WINDOWSHOWN = SDL_EVENT_WINDOW_SHOWN,
+    PGE_WINDOWHIDDEN = SDL_EVENT_WINDOW_HIDDEN,
+    PGE_WINDOWEXPOSED = SDL_EVENT_WINDOW_EXPOSED,
+    PGE_WINDOWMOVED = SDL_EVENT_WINDOW_MOVED,
+    PGE_WINDOWRESIZED = SDL_EVENT_WINDOW_RESIZED,
+    PGE_WINDOWSIZECHANGED = SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED,
+    PGE_WINDOWMINIMIZED = SDL_EVENT_WINDOW_MINIMIZED,
+    PGE_WINDOWMAXIMIZED = SDL_EVENT_WINDOW_MAXIMIZED,
+    PGE_WINDOWRESTORED = SDL_EVENT_WINDOW_RESTORED,
+    PGE_WINDOWENTER = SDL_EVENT_WINDOW_MOUSE_ENTER,
+    PGE_WINDOWLEAVE = SDL_EVENT_WINDOW_MOUSE_LEAVE,
+    PGE_WINDOWFOCUSGAINED = SDL_EVENT_WINDOW_FOCUS_GAINED,
+    PGE_WINDOWFOCUSLOST = SDL_EVENT_WINDOW_FOCUS_LOST,
+    PGE_WINDOWCLOSE = SDL_EVENT_WINDOW_CLOSE_REQUESTED,
+    PGE_WINDOWTAKEFOCUS = -1, /* No SDL3 equivalent */
+    PGE_WINDOWHITTEST = SDL_EVENT_WINDOW_HIT_TEST,
+    PGE_WINDOWICCPROFCHANGED = SDL_EVENT_WINDOW_ICCPROF_CHANGED,
+    PGE_WINDOWDISPLAYCHANGED = SDL_EVENT_WINDOW_DISPLAY_CHANGED,
+} PygameWindowEventCode;
+/*
+TODO: expose these window events in pygame API
+    SDL_EVENT_WINDOW_METAL_VIEW_RESIZED,
+    SDL_EVENT_WINDOW_DISPLAY_SCALE_CHANGED,
+    SDL_EVENT_WINDOW_SAFE_AREA_CHANGED,
+    SDL_EVENT_WINDOW_OCCLUDED,
+    SDL_EVENT_WINDOW_ENTER_FULLSCREEN,
+    SDL_EVENT_WINDOW_LEAVE_FULLSCREEN,
+    SDL_EVENT_WINDOW_DESTROYED,
+    SDL_EVENT_WINDOW_HDR_STATE_CHANGED,
+*/
+#endif
+
 typedef enum {
     /* Any SDL_* events here are for backward compatibility. */
     SDL_NOEVENT = 0,
@@ -465,6 +504,9 @@ typedef enum {
     PGE_MIDIIN,
     PGE_MIDIOUT,
 
+/* These PGE events are only needed on SDL2: SDL3 has dedicated events for
+ * these */
+#if !SDL_VERSION_ATLEAST(3, 0, 0)
     /* DO NOT CHANGE THE ORDER OF EVENTS HERE */
     PGE_WINDOWSHOWN,
     PGE_WINDOWHIDDEN,
@@ -484,6 +526,7 @@ typedef enum {
     PGE_WINDOWHITTEST,
     PGE_WINDOWICCPROFCHANGED,
     PGE_WINDOWDISPLAYCHANGED,
+#endif
 
     /* Here we define PGPOST_* events, events that act as a one-to-one
      * proxy for SDL events (and some extra events too!), the proxy is used
@@ -519,10 +562,8 @@ typedef enum {
     PGPOST_CONTROLLERTOUCHPADMOTION,
     PGPOST_CONTROLLERTOUCHPADUP,
     PGPOST_CONTROLLERSENSORUPDATE,
-#if !SDL_VERSION_ATLEAST(3, 0, 0)
     PGPOST_DOLLARGESTURE,
     PGPOST_DOLLARRECORD,
-#endif
     PGPOST_DROPFILE,
     PGPOST_DROPTEXT,
     PGPOST_DROPBEGIN,
@@ -547,9 +588,7 @@ typedef enum {
     PGPOST_MOUSEBUTTONDOWN,
     PGPOST_MOUSEBUTTONUP,
     PGPOST_MOUSEWHEEL,
-#if !SDL_VERSION_ATLEAST(3, 0, 0)
     PGPOST_MULTIGESTURE,
-#endif
     PGPOST_NOEVENT,
     PGPOST_QUIT,
     PGPOST_RENDER_TARGETS_RESET,
