@@ -260,7 +260,7 @@ MODINIT_DEFINE(constants)
     DEC_CONST(MOUSEBUTTONDOWN);
     DEC_CONST(MOUSEBUTTONUP);
     DEC_CONST(JOYAXISMOTION);
-    DEC_CONSTS(JOYBALLMOTION, PG_JOYBALLMOTION);
+    DEC_CONST(JOYBALLMOTION);
     DEC_CONST(JOYHATMOTION);
     DEC_CONST(JOYBUTTONDOWN);
     DEC_CONST(JOYBUTTONUP);
@@ -302,7 +302,7 @@ MODINIT_DEFINE(constants)
     DEC_CONST(FINGERMOTION);
     DEC_CONST(FINGERDOWN);
     DEC_CONST(FINGERUP);
-    DEC_CONSTS(MULTIGESTURE, PG_MULTIGESTURE);
+    DEC_CONST(MULTIGESTURE);
     DEC_CONST(AUDIODEVICEADDED);
     DEC_CONST(AUDIODEVICEREMOVED);
     DEC_CONST(MOUSEWHEEL);
