@@ -1687,7 +1687,7 @@ SaveTGA_RW(SDL_Surface *surface, SDL_RWops *out, int rle)
     int alpha = 0;
     struct TGAheader h;
     int srcbpp;
-    Uint8 surf_alpha;
+    SDL_BlendMode surf_blendmode;
     int have_surf_colorkey = 0;
     Uint32 surf_colorkey;
     SDL_Rect r;
@@ -1718,7 +1718,10 @@ SaveTGA_RW(SDL_Surface *surface, SDL_RWops *out, int rle)
     SDL_PixelFormatEnum output_format;
 #endif
 
-    SDL_GetSurfaceAlphaMod(surface, &surf_alpha);
+    if (!PG_GetSurfaceBlendMode(surface, &surf_blendmode)) {
+        return -1;
+    }
+
     if ((have_surf_colorkey = SDL_HasColorKey(surface))) {
         SDL_GetColorKey(surface, &surf_colorkey);
     }
@@ -1805,11 +1808,9 @@ SaveTGA_RW(SDL_Surface *surface, SDL_RWops *out, int rle)
         }
     }
 
-    /* Temporarily remove colorkey and alpha from surface so copies are
-       opaque */
-    SDL_SetSurfaceAlphaMod(surface, SDL_ALPHA_OPAQUE);
-    if (have_surf_colorkey) {
-        SDL_SetColorKey(surface, SDL_FALSE, surf_colorkey);
+    /* Temporarily set SDL_BLENDMODE_NONE so that copies are opaque */
+    if (!PG_SetSurfaceBlendMode(surface, SDL_BLENDMODE_NONE)) {
+        goto error;
     }
 
     r.x = 0;
@@ -1841,9 +1842,8 @@ SaveTGA_RW(SDL_Surface *surface, SDL_RWops *out, int rle)
     }
 
     /* restore flags */
-    SDL_SetSurfaceAlphaMod(surface, surf_alpha);
-    if (have_surf_colorkey) {
-        SDL_SetColorKey(surface, SDL_TRUE, surf_colorkey);
+    if (!PG_SetSurfaceBlendMode(surface, surf_blendmode)) {
+        goto error;
     }
 
     free(rlebuf);
